backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.2.0
numpy==1.26.3
openai==1.10.0
python-multipart==0.0.6
python-pptx==0.6.23
matplotlib==3.8.2
pytest==7.4.4
httpx==0.26.0
